+++ CHANGES.rst
@@ -4,6 +4,8 @@
 * Bug fix: Name collisions with version_helper.py
 * Bug fix: Subpath glob matching always respects case
 * Feature: @argify works on view classes
+* Feature: @argify can inject types that consume multiple parameters
+* Feature: Parameter types can be a dotted path
 
 0.1.0
 -----